+++ src/main/java/ecse321/fall2014/group3/bomberman/physics/entity/Player.java
@@ -9,12 +9,11 @@
 package ecse321.fall2014.group3.bomberman.physics.entity;
 
 import com.flowpowered.math.vector.Vector2f;
-import com.flowpowered.math.vector.Vector2i;
 
 import ecse321.fall2014.group3.bomberman.nterface.SpriteInfo;
 
 public class Player extends Entity {
-    private static final SpriteInfo STILL_SPRITE = new SpriteInfo("main", 0, Vector2i.ONE);
+    private static final SpriteInfo STILL_SPRITE = new SpriteInfo("main", 0, Vector2f.ONE);
     private volatile int blastRadius = 1;
     private volatile int placementBombCount = 1;
     private volatile float speed = 1;
